Finished
After the first day, I created the background colors, worked on the nav and body elements, creating background colors and fonts. I also started work on the pseudo classes. On the second day, I finished on the pseudo classes, set the days to alternate between colors, and I started and finished the block-quote element, through adding color and paddings, etc.